Remove double as float feature
This commit removes the feature of using double-based
types as if they were float-based. This feature was confusing
and prevented the use of real double-based types.

+++ src/Libs/FShade.Core/Serializer.fs
@@ -1859,15 +1859,15 @@ module Serializer =
                 let maxLod = 
                     match src.ReadByte() with
                     | 0uy -> None
-                    | _ -> Some (src.ReadDouble())
+                    | _ -> Some (src.ReadSingle())
                 let minLod = 
                     match src.ReadByte() with
                     | 0uy -> None
-                    | _ -> Some (src.ReadDouble())
+                    | _ -> Some (src.ReadSingle())
                 let mipLodBias = 
                     match src.ReadByte() with
                     | 0uy -> None
-                    | _ -> Some (src.ReadDouble())
+                    | _ -> Some (src.ReadSingle())
 
                 {
                     AddressU = addressU
